requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
cycler==0.10.0
h11==0.12.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
outcome==1.1.0
Pillow==8.4.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
scipy==1.7.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
